Front end with handlebars

+++ src/controllers/coops.js
@@ -8,17 +8,13 @@ class Coops {
 
 	// creating a function for add cooperative
 	static async createCoop(req,res){
-		
 		// // try and catch to find if not exist and create new coop
 		try{ 
 			const {coopName,coopLocation,tin} = req.body;
-			const UserId= req.user.id;
+			const UserId= req.user.userFind.id;
 			const Op = Sequelize.Op;
 		// 	// find if exist
-			const findOne = await Coop.findOne({where: {
-				[Op.or]:[{tin},{UserId}]
-			}
-			});
+			const findOne = await Coop.findOne({where: {tin}});
 			if (!findOne) {
 				const coudinary_links = await cloud(req.files);
 				const createcoop = Coop.create({
@@ -32,26 +28,20 @@ class Coops {
 					coopSignL:coudinary_links[3],
 					leaderCert:coudinary_links[4]
 				}).then(result =>{
-					return res.status(200).send({
-						status:res.statusCode,
+					return res.render('add-coop',{
+						user:req.user.userFind,
+						role:{
+							isEax:req.user.role.isEax(req.user.userFind),
+							isCoop:req.user.role.isCoop(req.user.userFind),
+							isBidder:req.user.role.isBidder(req.user.userFind),
+						},
 						message:'cooperative created successfully!',
-						result
 					})
 				})
-			}else{
-				return res.status(500).send({
-					status:res.statusCode,
-					message:'something went wrong!'
-					
-				})
 			}
-			
 		}
 		catch(err){
-			res.status(500).send({
-				status:res.statusCode,
-				message:'Check internet connection!',
-			})
+			return res.render('500');
 		}
 	}
 	// update coop function
@@ -87,46 +77,86 @@ class Coops {
 	static async deleteCoop(req,res){
 		// check if tin is not empty
 		try{
-			const{tin}=req.body;
+			const {tin,coopName}= req.params;
+			const findOne = await Coop.findOne({where:{tin,coopName}});
+			if (findOne) {
 				await Coop.destroy({where:{tin}}).then(result=>{
 					if (result) {
-						return res.status(200).send({
-							status:res.statusCode,
+						return res.render('del-coop',{
+							user:req.user.userFind,
+							role:{
+								isEax:req.user.role.isEax(req.user.userFind),
+								isCoop:req.user.role.isCoop(req.user.userFind),
+								isBidder:req.user.role.isBidder(req.user.userFind),
+							},
 							message:'cooperative has been deleted successfully!',
-						})
-					}else{
-						res.status(500).send({
-							status:res.statusCode,
-							message:'cooperative with that tin not exist yet!'
 						})
 					}
 				})
+			} else{
+				return res.render('del-coop',{
+					user:req.user.userFind,
+					role:{
+						isEax:req.user.role.isEax(req.user.userFind),
+						isCoop:req.user.role.isCoop(req.user.userFind),
+						isBidder:req.user.role.isBidder(req.user.userFind),
+					},
+					message:`cooperative with tin ${tin} not exist !`,
+				})
+			}
+				
 			
 		}catch(err){
-			return res.status(500).send({
-				status:res.statusCode,
-				message:'something went wrong on server',
-			})
+			return res.render('500');
 		}
 	}
 	// get all coops
 	static async getCoops(req,res){
 		try{	
 			await Coop.findAll().then(result =>{
-				return res.status(200).send({
-					status:res.statusCode,
-					message:'All cooperatives fetched successfully!',
+				return res.render('all-coops',{
 					result,
+					user:req.user.userFind,
+					role:{
+						isEax:req.user.role.isEax(req.user.userFind),
+						isCoop:req.user.role.isCoop(req.user.userFind),
+						isBidder:req.user.role.isBidder(req.user.userFind),
+					},
 				})
 			})
 		}catch(err){
-			res.status(500).send({
-				status:res.statusCode,
-				message:'something went wrong on server!'
-			})
+			res.render('500');
 		}
 	}
-	// 
+	// get info of coop to delete
+	static async getDelInfo(req,res){
+		try{
+			const{coopName}=req.body;
+			const findOne = await Coop.findOne({where:{coopName}});
+			if (findOne != null) {
+				return res.render('del-coop',{
+					findOne,
+					user:req.user.userFind,
+					role:{
+						isEax:req.user.role.isEax(req.user.userFind),
+						isCoop:req.user.role.isCoop(req.user.userFind),
+						isBidder:req.user.role.isBidder(req.user.userFind),
+					},
+					message:`${coopName}  found  successfully!`
+				})
+			} 
+			return res.render('del-coop',{
+				user:req.user.userFind,
+				role:{
+					isEax:req.user.role.isEax(req.user.userFind),
+					isCoop:req.user.role.isCoop(req.user.userFind),
+					isBidder:req.user.role.isBidder(req.user.userFind),
+				},
+			})
+		}catch(err){
+			return res.render('500');
+		}
+	}
 }
 
 // export coops
